Split functions into seperate files
